Made JSON changes to resumeData.js to modularize the way data is being presented in the Resume.jsx route. All data is being passed into the Resume component so that when backend database is implemented, it will be able to handle rendering the page in the same fashion.
Used React-Bootstrap and Tachyons for the front-end components. Need to rearrange the grid so that the containers do not resize strangely. Created experience component to render in bootstrap collapsable work history dynamically based on experience prop that contains an array of all data being passed. Animations are not rendering smoothly when collapsing or changing tabs. Particles are not rendering correctly on the left and right portions of the body context.

Implemented DevBlog: No functionality other than providing information on the progress of the web application. DevBlogData is being passed into DevBlog component from JSON.

Implemented Projects: Page that displays my projects with github links. Link bound to github logo, may need to be more specific on the link. It may be hard for users to notice the link.