Created a landing page that describes the page being under construction. OOF! Added React-Bootstrap version of Carousel, replaced the old one.
Added context to the Contact Me page, simple image with some contact information. Will replace this with form and email transaction API as described in Contact Me

Will likely be switching gears to .Net to prepare for interviews, and will return to finish up the front end and begin working on the Node.js service

Deployed Web App as is for now.